project-2-medium/backend/auth_server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==44.0.0
python-multipart==0.0.20
pydantic==2.10.4
